added ML project
The proposed model work on the 
Convolutional Neural networks(CNN) to extract the 
spatial features from the input signals and a 
Bidirectional Long Short-Term Memory (BiLSTM) 
network to capture the time dependencies in the 
signals. this model mainly focuses on the variations in the non-
sationary signals to predict the disruptions even before 
they occur.